projects/20260109_AIエージェント開発/requirements.txt
openai
requests
gitpython
tweepy
python-dotenv
